update specs, there are problems in the old one
